==> codebase/07_replication_assess_match_quality/08_bunmd_birthplace_coverage.Rmd <==
---
title: "BUNMD Birthplace Coverage" 
author: "Casey Breen" 
---

Summary: BUNMD Birthplace Coverage. 

```{r}
## Library packages 
library(tidyverse)
library(data.table)
library(censocdev)
library(cowplot)
library(gt)
library(ipumsr)
library(here)
library(janitor)
```



```{r}
bunmd <- fread("/censoc/data/censoc_v2/bunmd_v2.csv")
```


```{r}
bunmd_tab <- bunmd %>% 
  group_by(byear, sex) %>% 
  summarize(bpl_available = mean(!is.na(bpl)))
```


```{r}
bunmd_bpl_plot <- bunmd_tab %>% 
  mutate(sex = case_when(
    sex == 1 ~ "Men",
    sex == 2 ~ "Women"
  )) %>% 
  filter(byear %in% 1900:1940) %>% 
  filter(!is.na(sex)) %>% 
  ggplot(aes(x = byear, y = bpl_available, shape = sex, linetype = sex)) + 
  geom_line() + 
  geom_point() + 
  theme_cowplot() + 
  ylim(0, 1) + 
  labs(x = "Birth Year",
       y = "Proportion Birthplace Available") + 
  theme(legend.position = "bottom", legend.title = element_blank()) 


ggsave(bunmd_bpl_plot, filename = here("vignettes/assess_match_quality/figs/bunmd_bpl_plot.pdf"), height = 5, width = 7)
```
